+++ .gitignore
@@ -17,70 +17,3 @@
 *.suo
 *.log
 *.cache
-/UniversitySchedule/Infrastructure/obj/Debug/net8.0/Infrastructure.AssemblyInfo.cs
-/UniversitySchedule/Infrastructure/obj/Debug/net8.0/Infrastructure.AssemblyInfoInputs.cache
-/UniversitySchedule/Infrastructure/obj/Debug/net8.0/Infrastructure.assets.cache
-/UniversitySchedule/Infrastructure/obj/project.assets.json
-/UniversitySchedule/Infrastructure/obj/Infrastructure.csproj.nuget.g.props
-/UniversitySchedule/Infrastructure/obj/Infrastructure.csproj.nuget.dgspec.json
-/UniversitySchedule/Infrastructure/obj/project.nuget.cache
-/UniversitySchedule/Infrastructure/obj/Debug/net8.0/Infrastructure.csproj.AssemblyReference.cache
-/UniversitySchedule/Infrastructure/obj/Debug/net8.0/Infrastructure.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Domain/obj/Debug/net8.0/Domain.AssemblyInfoInputs.cache
-/UniversitySchedule/.vs/UniversitySchedule/v17/.suo
-/UniversitySchedule/.vs/UniversitySchedule/v17/DocumentLayout.json
-/UniversitySchedule/.vs/UniversitySchedule/DesignTimeBuild/.dtbcache.v2
-/UniversitySchedule/Domain/obj/Debug/net8.0/Domain.assets.cache
-/UniversitySchedule/Domain/obj/Debug/net8.0/Domain.AssemblyInfo.cs
-/UniversitySchedule/Domain/obj/Debug/net8.0/Domain.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Domain/obj/Domain.csproj.nuget.dgspec.json
-/UniversitySchedule/Domain/obj/Domain.csproj.nuget.g.props
-/UniversitySchedule/Domain/obj/project.assets.json
-/UniversitySchedule/Domain/obj/project.nuget.cache
-/UniversitySchedule/Application/obj/project.assets.json
-/UniversitySchedule/Application/obj/project.nuget.cache
-/UniversitySchedule/Application/obj/Application.csproj.nuget.g.props
-/UniversitySchedule/Application/obj/Debug/net8.0/Application.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Application/obj/Debug/net8.0/Application.assets.cache
-/UniversitySchedule/Application/obj/Debug/net8.0/Application.csproj.AssemblyReference.cache
-/UniversitySchedule/Application/obj/Debug/net8.0/Application.AssemblyInfoInputs.cache
-/UniversitySchedule/Application/obj/Debug/net8.0/Application.AssemblyInfo.cs
-/UniversitySchedule/Application/obj/Application.csproj.nuget.dgspec.json
-/UniversitySchedule/Infrastructure/obj/
-/UniversitySchedule/Domain/obj/
-/UniversitySchedule/Application/obj
-/UniversitySchedule/UniversitySchedule/obj/Debug
-/UniversitySchedule/UniversitySchedule/obj/Debug/UniversitySchedule.csproj.AssemblyReference.cache
-/UniversitySchedule/UniversitySchedule/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.GlobalUsings.g.cs
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.assets.cache
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.AssemblyInfoInputs.cache
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.AssemblyInfo.cs
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.GlobalUsings.g.cs
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.assets.cache
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.AssemblyInfoInputs.cache
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.AssemblyInfo.cs
-/UniversitySchedule/Application/obj/Release/net8.0/Application.GlobalUsings.g.cs
-/UniversitySchedule/Application/obj/Release/net8.0/Application.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Application/obj/Release/net8.0/Application.assets.cache
-/UniversitySchedule/Application/obj/Release/net8.0/Application.AssemblyInfoInputs.cache
-/UniversitySchedule/Application/obj/Release/net8.0/Application.AssemblyInfo.cs
-/UniversitySchedule/UniversitySchedule/obj/Debug/UniversitySchedule.csproj.AssemblyReference.cache
-/UniversitySchedule/UniversitySchedule/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.GlobalUsings.g.cs
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.assets.cache
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.AssemblyInfoInputs.cache
-/UniversitySchedule/Infrastructure/obj/Release/net8.0/Infrastructure.AssemblyInfo.cs
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.GlobalUsings.g.cs
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.assets.cache
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.AssemblyInfoInputs.cache
-/UniversitySchedule/Domain/obj/Release/net8.0/Domain.AssemblyInfo.cs
-/UniversitySchedule/Application/obj/Release/net8.0/Application.GlobalUsings.g.cs
-/UniversitySchedule/Application/obj/Release/net8.0/Application.GeneratedMSBuildEditorConfig.editorconfig
-/UniversitySchedule/Application/obj/Release/net8.0/Application.assets.cache
-/UniversitySchedule/Application/obj/Release/net8.0/Application.AssemblyInfoInputs.cache
-/UniversitySchedule/Application/obj/Release/net8.0/Application.AssemblyInfo.cs
