--- UniversitySchedule/UniversitySchedule.sln
@@ -1,44 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.13.35825.156
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{36D591C7-65C7-A0D1-1CBC-10CDE441BDC8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "Application\Application.csproj", "{26448D8D-F44D-4468-833D-3E6B2619230A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{A1921C07-259F-43DF-B7F1-2FE6355096BE}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{EA6C1670-349E-4A03-9962-A76B5B06E27D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{26448D8D-F44D-4468-833D-3E6B2619230A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{26448D8D-F44D-4468-833D-3E6B2619230A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{26448D8D-F44D-4468-833D-3E6B2619230A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{26448D8D-F44D-4468-833D-3E6B2619230A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A1921C07-259F-43DF-B7F1-2FE6355096BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A1921C07-259F-43DF-B7F1-2FE6355096BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A1921C07-259F-43DF-B7F1-2FE6355096BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A1921C07-259F-43DF-B7F1-2FE6355096BE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA6C1670-349E-4A03-9962-A76B5B06E27D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA6C1670-349E-4A03-9962-A76B5B06E27D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA6C1670-349E-4A03-9962-A76B5B06E27D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA6C1670-349E-4A03-9962-A76B5B06E27D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{26448D8D-F44D-4468-833D-3E6B2619230A} = {36D591C7-65C7-A0D1-1CBC-10CDE441BDC8}
-		{A1921C07-259F-43DF-B7F1-2FE6355096BE} = {36D591C7-65C7-A0D1-1CBC-10CDE441BDC8}
-		{EA6C1670-349E-4A03-9962-A76B5B06E27D} = {36D591C7-65C7-A0D1-1CBC-10CDE441BDC8}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FB527CF5-B5D6-4BD4-9E99-27BA3796C7A0}
-	EndGlobalSection
-EndGlobal
