+++ src/Application/Services/IAuditoriumService.cs
