+++ src/Domain/Entities/AcademicBuilding.cs
