+++ src/Domain/Entities/Auditorium.cs
