+++ src/Domain/Entities/ScheduleItems.cs
