+++ src/Domain/Enums/EducationFormEnum.cs
