+++ src/Domain/Enums/EducationLevelEnum.cs
