+++ src/Domain/Enums/LessonTypeEnum.cs
