+++ src/Domain/Enums/UserTypeEnum.cs
