+++ src/Infrastructure/Repositories/AuditoriumRepository.cs
