+++ src/Infrastructure/Repositories/IAuditoriumRepository.cs
