+++ src/Infrastructure/Repositories/ScheduleItemsRepository.cs
